+++ packages/core/src/index.ts
@@ -3,6 +3,7 @@ export * from './utils/useInjectStyles';
 export * from './utils/staggerStyle';
 export * from './utils/usePageBackground';
 export * from './utils/useDisclosure';
+export * from './utils/slot';
 
 // Tokens
 export * from './tokens';